Feat: User and  userReport CRUD

+++ src/services/userReport/deleteUserReport.js
@@ -3,10 +3,16 @@ import UserReportRepository from '../../domain/userReport/userReport.repository.
 export async function deleteUserReport (req, res) {
   const userReportRepository = new UserReportRepository()
 
-  const user = await userReportRepository.deleteUserReport(parseInt(req.params.id))
+  const userReportFound = await userReportRepository.getUserReport(parseInt(req.params.id))
 
-  if (user) {
-    return res.status(200).json(user)
+  if (!userReportFound) {
+    return res.status(404).json({ message: 'No such Report Found' })
+  } else {
+    const userReport = await userReportRepository.deleteUserReport(parseInt(req.params.id))
+    if (userReport) {
+      return res.status(200).json(userReport)
+    } else {
+      return res.status(500).json({ message: 'Someting Went Wrong While deleting user Report' })
+    }
   }
-  res.status(404).json({ message: 'No such Report Found' })
 }
